Run the modern (calibration) model

Run the modern (calibration) model

Usage

run_modern(
  modern_elevation = NULL,
  modern_species = NULL,
  dx = 0.1,
  ChainNums = seq(1, 3),
  n.iter = 40000,
  n.burnin = 10000,
  n.thin = 15,
  validation.run = FALSE,
  sigma_z_priors = NULL,
  scale_x = FALSE,
  fold = 1
)

Arguments

modern_elevation

A dataframe of modern elevations

modern_species

A dataframe of modern counts (to be sorted with sort_modern)

dx

The elevation interval for spacing the spline knots. Defaults to 0.2.

ChainNums

The number of MCMC chains to run

n.iter

The number of iterations

n.burnin

The number of burnin samples

n.thin

The number of thinning

validation.run

Set to TRUE if running a validation

sigma_z_priors

provide priors for foram variability if available

scale_x

Set to TRUE to scale elevation data to have mean 0 and sd 1

fold

Fold number for CV

Value

Nothing is returned, the relevant data will be saved.

Examples

test_modern_mod <- run_modern(
  modern_elevation = NJ_modern_elevation,
  modern_species = NJ_modern_species,
  n.iter = 10,
  n.burnin = 1,
  n.thin = 1
)
#> Start chain ID  1 
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 175
#>    Unobserved stochastic nodes: 3136
#>    Total graph size: 20460
#> 
#> Initializing model
#> 
#> Hooraah, Chain 1 has finished! 
#> Start chain ID  2 
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 175
#>    Unobserved stochastic nodes: 3136
#>    Total graph size: 20460
#> 
#> Initializing model
#> 
#> Hooraah, Chain 2 has finished! 
#> Start chain ID  3 
#> Compiling model graph
#>    Resolving undeclared variables
#>    Allocating nodes
#> Graph information:
#>    Observed stochastic nodes: 175
#>    Unobserved stochastic nodes: 3136
#>    Total graph size: 20460
#> 
#> Initializing model
#> 
#> Hooraah, Chain 3 has finished! 
#> Run complete!